Simulate one BF-BOIN trial

Simulate one BF-BOIN trial

Usage

sim.one.trial(
  trial.id = 1,
  target = 0.25,
  p.true = c(0.1, 0.3, 0.5),
  ncohort = 10,
  cohortsize = 3,
  n.earlystop = 100,
  startdose = 1,
  titration = FALSE,
  p.saf = 0.6 * target,
  p.tox = 1.4 * target,
  cutoff.eli = 0.95,
  extrasafe = FALSE,
  offset = 0.05,
  boundMTD = FALSE,
  n.cap = 12,
  end.backfill = TRUE,
  n.per.month = 3,
  dlt.window = 1,
  p.response.true = c(1, 1, 1),
  three.plus.three = FALSE,
  accrual = "uniform",
  backfill.assign = "highest"
)

Arguments

trial.id

an ID for the trial

target

the target DLT rate

p.true

a vector containing the true toxicity probabilities of the investigational dose levels.

ncohort

the total number of cohorts

cohortsize

the cohort size

n.earlystop

the early stopping parameter. If the number of patients treated at the current dose reaches n.earlystop, stop the trial and select the MTD based on the observed data. The default value n.earlystop=100 essentially turns off this type of early stopping.

startdose

the starting dose level for the trial

titration

set titration=TRUE to perform dose escalation with cohort size = 1 to accelerate dose escalation at the begining of the trial.

p.saf

the highest toxicity probability that is deemed subtherapeutic (i.e. below the MTD) such that dose escalation should be undertaken. The default value is p.saf=0.6*target.

p.tox

the lowest toxicity probability that is deemed overly toxic such that deescalation is required. The default value is p.tox=1.4*target).

cutoff.eli

the cutoff to eliminate an overly toxic dose for safety. We recommend the default value of (cutoff.eli=0.95) for general use.

extrasafe

set extrasafe=TRUE to impose a more stringent stopping rule

offset

a small positive number (between 0 and 0.5) to control how strict the stopping rule is when extrasafe=TRUE. A larger value leads to a more strict stopping rule. The default value offset=0.05 generally works well.

boundMTD

set boundMTD=TRUE to impose the condition: the isotonic estimate of toxicity probability for the selected MTD must be less than de-escalation boundary.

n.cap

permanently close a dose for backfilling if the number of patients assigned to the dose reaches n.cap

end.backfill

when the dose escalation ends, the backfilling by definition also ends. Default is TRUE.

n.per.month

patient accrual rate per month

dlt.window

DLT assessment window (months)

p.response.true

a vector containing the true response probabilities of the investigational dose levels

three.plus.three

modify the decision from de-escalation to stay when observing 1 DLT out of 3 patients

accrual

"uniform" or "poisson", according to whether accrual distribution is uniform (consistent with Shiny App) or a Poisson process (consistent with publication)

backfill.assign

How to assign backfill dose given the open backfill doses. Options are "highest" (default), "lowest", or "random".

Value

A data frame with the number of patients and number of DLTs at each dose level